renamed reservation creator to reservation.rb
