+++ lib/reservation.rb
