+++ spec/spec_helper.rb
@@ -10,5 +10,5 @@ require 'awesome_print'
 Minitest::Reporters.use! Minitest::Reporters::SpecReporter.new
 
 # Require_relative your lib files here!
-require_relative '../lib/reservation_creator'
+require_relative '../lib/reservation'
 require_relative '../lib/booking_system'
